move shortcut to api
